tccelf: Experiment with emitting debug symbols to gdb at runtime
gdb and lldb have a somewhat standard JIT interface that allows a
program to emit debug symbols at runtime by providing an object image
and then calling a special hook that the debugger then intercepts.

Enable JIT debugging in gdb with:
set debug jit on

Given that tinycc can now emit dwarf-5 debug information, I figured I
would attempt to hook this interface up to allow gdb and lldb to follow
JIT-compiled code.
This implementation is far from optimal, I'm not very familiar with the
TCC codebase. I just duplicated some of the ELF file saving code to
store a full elf image to an in-memory buffer with open_memstream(3).

This currently does not fully work yet. It is emitting the object image,
and gdb accepts it, but I have a suspicion that tcc may be emitting
invalid dwarf information for some reason. This is possibly caused by my
changes, of course.